feat: optimized DispatcherSelectorSpec and additional case (#31862)
