+++ akka-actor-typed-tests/src/test/scala/akka/actor/typed/scaladsl/DispatcherSelectorSpec.scala
@@ -8,119 +8,156 @@ import com.typesafe.config.Config
 import com.typesafe.config.ConfigFactory
 import org.scalatest.wordspec.AnyWordSpecLike
 
+import akka.actor.ActorCell
 import akka.actor.BootstrapSetup
 import akka.actor.setup.ActorSystemSetup
 import akka.actor.testkit.typed.scaladsl.ActorTestKit
 import akka.actor.testkit.typed.scaladsl.LogCapturing
 import akka.actor.testkit.typed.scaladsl.LoggingTestKit
 import akka.actor.testkit.typed.scaladsl.ScalaTestWithActorTestKit
-import akka.actor.testkit.typed.scaladsl.TestProbe
 import akka.actor.typed.ActorRef
 import akka.actor.typed.ActorSystem
 import akka.actor.typed.Behavior
+import akka.actor.typed.DispatcherSelector
 import akka.actor.typed.Props
 import akka.actor.typed.SpawnProtocol
+import akka.actor.typed.internal.adapter.ActorContextAdapter
+import akka.actor.typed.scaladsl.AskPattern._
+import akka.dispatch.BoundedMessageQueueSemantics
+import akka.dispatch.BoundedNodeMessageQueue
+import akka.dispatch.Dispatchers
+import akka.dispatch.MessageQueue
+import akka.dispatch.NodeMessageQueue
 
 object DispatcherSelectorSpec {
   val config = ConfigFactory.parseString("""
-      ping-pong-dispatcher {
+      specific-dispatcher {
         executor = thread-pool-executor
         type = PinnedDispatcher
       }
-    """)
-
-  object PingPong {
-    case class Ping(replyTo: ActorRef[Pong])
-    case class Pong(threadName: String)
-
-    def apply(): Behavior[Ping] =
-      Behaviors.receiveMessage[Ping] { message =>
-        message.replyTo ! Pong(Thread.currentThread().getName)
-        Behaviors.same
+      specific-mailbox {
+        mailbox-type = "akka.dispatch.NonBlockingBoundedMailbox"
+        mailbox-capacity = 4
       }
-
-  }
-
+      specific-dispatcher-with-mailbox {
+        type = Dispatcher
+        executor = "thread-pool-executor"
+        mailbox-type = "akka.dispatch.NonBlockingBoundedMailbox"
+        mailbox-capacity = 4
+      }
+    """)
 }
 
 class DispatcherSelectorSpec(config: Config)
     extends ScalaTestWithActorTestKit(config)
     with AnyWordSpecLike
     with LogCapturing {
-  import DispatcherSelectorSpec.PingPong
-  import DispatcherSelectorSpec.PingPong._
 
   def this() = this(DispatcherSelectorSpec.config)
 
+  case class WhatsYourDispatcherAndMailbox(replyTo: ActorRef[(String, MessageQueue)])
+
+  private def behavior: Behavior[WhatsYourDispatcherAndMailbox] =
+    Behaviors.setup { context =>
+      Behaviors.receiveMessage[WhatsYourDispatcherAndMailbox] {
+        case WhatsYourDispatcherAndMailbox(replyTo) =>
+          val result = context match {
+            case adapter: ActorContextAdapter[_] =>
+              adapter.classicActorContext match {
+                case cell: ActorCell =>
+                  (cell.dispatcher.id, cell.mailbox.messageQueue)
+                case unexpected => throw new RuntimeException(s"Unexpected: $unexpected")
+              }
+            case unexpected => throw new RuntimeException(s"Unexpected: $unexpected")
+          }
+
+          replyTo ! result
+          Behaviors.stopped
+      }
+    }
+
   "DispatcherSelector" must {
 
-    "select dispatcher from config" in {
-      val probe = createTestProbe[Pong]()
-      val pingPong = spawn(PingPong(), Props.empty.withDispatcherFromConfig("ping-pong-dispatcher"))
-      pingPong ! Ping(probe.ref)
+    "default to unbounded" in {
+      val actor = spawn(behavior)
+      val (_, mailbox) = actor.ask(WhatsYourDispatcherAndMailbox).futureValue
+      mailbox shouldBe a[NodeMessageQueue]
+    }
 
-      val response = probe.receiveMessage()
-      response.threadName should startWith("DispatcherSelectorSpec-ping-pong-dispatcher")
+    "select a blocking dispatcher" in {
+      val actor = spawn(behavior, DispatcherSelector.blocking())
+      val (dispatcherId, _) = actor.ask(WhatsYourDispatcherAndMailbox).futureValue
+      dispatcherId shouldBe Dispatchers.DefaultBlockingDispatcherId
+    }
+
+    "select dispatcher from config" in {
+      val actor = spawn(behavior, Props.empty.withDispatcherFromConfig("specific-dispatcher"))
+      val (dispatcherId, _) = actor.ask(WhatsYourDispatcherAndMailbox).futureValue
+      dispatcherId shouldBe "specific-dispatcher"
     }
 
     "detect unknown dispatcher from config" in {
-      val probe = createTestProbe[Pong]()
+      val probe = createTestProbe[(String, MessageQueue)]()
       LoggingTestKit.error("Spawn failed").expect {
-        val ref = spawn(PingPong(), Props.empty.withDispatcherFromConfig("unknown"))
+        val ref = spawn(behavior, Props.empty.withDispatcherFromConfig("unknown"))
         probe.expectTerminated(ref)
       }
     }
 
     "select same dispatcher as parent" in {
-      val parent = spawn(SpawnProtocol(), Props.empty.withDispatcherFromConfig("ping-pong-dispatcher"))
-      val childProbe = createTestProbe[ActorRef[Ping]]()
-      parent ! SpawnProtocol.Spawn(PingPong(), "child", Props.empty.withDispatcherSameAsParent, childProbe.ref)
+      val parent = spawn(SpawnProtocol(), Props.empty.withDispatcherFromConfig("specific-dispatcher"))
 
-      val probe = createTestProbe[Pong]()
-      val child = childProbe.receiveMessage()
-      child ! Ping(probe.ref)
+      val child = parent.ask { replyTo: ActorRef[ActorRef[WhatsYourDispatcherAndMailbox]] =>
+        SpawnProtocol.Spawn(behavior, "child", DispatcherSelector.sameAsParent(), replyTo)
+      }.futureValue
 
-      val response = probe.receiveMessage()
-      response.threadName should startWith("DispatcherSelectorSpec-ping-pong-dispatcher")
+      val (dispatcherId, _) = child.ask(WhatsYourDispatcherAndMailbox).futureValue
+      dispatcherId shouldBe "specific-dispatcher"
     }
 
     "select same dispatcher as parent, several levels" in {
-      val grandParent = spawn(SpawnProtocol(), Props.empty.withDispatcherFromConfig("ping-pong-dispatcher"))
-      val parentProbe = createTestProbe[ActorRef[SpawnProtocol.Spawn[Ping]]]()
-      grandParent ! SpawnProtocol.Spawn(
-        SpawnProtocol(),
-        "parent",
-        Props.empty.withDispatcherSameAsParent,
-        parentProbe.ref)
+      val guardian = spawn(SpawnProtocol(), Props.empty.withDispatcherFromConfig("specific-dispatcher"))
 
-      val childProbe = createTestProbe[ActorRef[Ping]]()
-      grandParent ! SpawnProtocol.Spawn(PingPong(), "child", Props.empty.withDispatcherSameAsParent, childProbe.ref)
+      val parent = guardian.ask { replyTo: ActorRef[ActorRef[SpawnProtocol.Command]] =>
+        SpawnProtocol.Spawn(SpawnProtocol(), "parent", DispatcherSelector.sameAsParent(), replyTo)
+      }.futureValue
 
-      val probe = createTestProbe[Pong]()
-      val child = childProbe.receiveMessage()
-      child ! Ping(probe.ref)
+      val child = parent.ask { replyTo: ActorRef[ActorRef[WhatsYourDispatcherAndMailbox]] =>
+        SpawnProtocol.Spawn(behavior, "child", DispatcherSelector.sameAsParent(), replyTo)
+      }.futureValue
 
-      val response = probe.receiveMessage()
-      response.threadName should startWith("DispatcherSelectorSpec-ping-pong-dispatcher")
+      val (dispatcherId, _) = child.ask(WhatsYourDispatcherAndMailbox).futureValue
+      dispatcherId shouldBe "specific-dispatcher"
     }
 
     "use default dispatcher if selecting parent dispatcher for user guardian" in {
       val sys = ActorSystem(
-        PingPong(),
+        behavior,
         "DispatcherSelectorSpec2",
         ActorSystemSetup.create(BootstrapSetup()),
-        Props.empty.withDispatcherSameAsParent)
+        DispatcherSelector.sameAsParent())
       try {
-        val probe = TestProbe[Pong]()(sys)
-        sys ! Ping(probe.ref)
-
-        val response = probe.receiveMessage()
-        response.threadName should startWith("DispatcherSelectorSpec2-akka.actor.default-dispatcher")
+        val (dispatcherId, _) = sys.ask(WhatsYourDispatcherAndMailbox).futureValue
+        dispatcherId shouldBe "akka.actor.default-dispatcher"
       } finally {
         ActorTestKit.shutdown(sys)
       }
     }
 
-  }
+    "select an specific dispatcher from config with mailbox selector" in {
+      val actor =
+        spawn(behavior, DispatcherSelector.fromConfig("specific-dispatcher").withMailboxFromConfig("specific-mailbox"))
+      val (dispatcherId, mailbox) = actor.ask(WhatsYourDispatcherAndMailbox).futureValue
+      mailbox shouldBe a[BoundedMessageQueueSemantics]
+      mailbox.asInstanceOf[BoundedNodeMessageQueue].capacity should ===(4)
+      dispatcherId shouldBe "specific-dispatcher"
+    }
 
+    "select an mailbox via dispatcher config would not work" in {
+      val actor = spawn(behavior, DispatcherSelector.fromConfig("specific-dispatcher-with-mailbox"))
+      val (dispatcherId, mailbox) = actor.ask(WhatsYourDispatcherAndMailbox).futureValue
+      mailbox shouldBe a[NodeMessageQueue]
+      dispatcherId shouldBe "specific-dispatcher-with-mailbox"
+    }
+  }
 }
